backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
psycopg2-binary>=2.9,<2.10
mozilla-django-oidc>=4.0,<4.1
django-filter>=24.0,<25.0
python-dateutil>=2.9,<2.10
gunicorn>=22.0,<23.0
whitenoise>=6.7,<6.8
